requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncpg>=0.28.0
# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
# Additional
bcrypt>=4.0.1
email-validator>=2.0.0
aiosqlite>=0.19.0
# Production
gunicorn>=21.2.0
